refactor : 공통 메소드 변경
- Utils 에서 CommonUtils 로 변경
- getQuestion() 메소드의 이동 : QuestionController 에서 CommonUtils 로
- validateUserId() 메소드의 이동 : Question 에서 User 로
